% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{benchmark_parameters}
\alias{benchmark_parameters}
\title{Formal benchmarking for sensitivity parameters}
\usage{
benchmark_parameters(
  weights,
  weights_benchmark,
  k_sigma = 1,
  k_rho = 1,
  Y,
  Z,
  sigma2,
  estimand = "ATT"
)
}
\arguments{
\item{weights}{Vector of estimated weights}

\item{weights_benchmark}{Vector of estimated weights, omitting the covariate (or set of covariates) being used to benchmark}

\item{k_sigma}{Relative ability of omitted confounder to explain variation in the true weights. If \code{k_sigma > 1}, then we expect the omitted confounder to explain more variation in the true weights than the benchmarked covariate(s). If \code{k_sigma < 1}, then we expect the omitted confounder to explain less of the variation in the true weights than the benchmarked covariate(s). Default is set to 1.}

\item{k_rho}{Relative correlation of omitted confounder with the outcome. If \code{k_rho > 1}, then we expect the omitted confounder to be more correlated with the outcome than the benchmarked covariate(s). If \code{k_rho < 1}, then we expect the omitted confounder to be less correlated with the outcome than the benchmarked covariate(s). Default is set to 1.}

\item{Y}{Vector of outcomes}

\item{Z}{Vector of treatment assignment (Only necessary if \code{estimand = "PATE"} in order to estimate covariances)}

\item{sigma2}{Estimated variance of the outcome (i.e., var(Y) for obervational setting; var(tau) for generalization setting)}

\item{estimand}{String specifying the estimand of interest. Valid inputs are "PATE", "Augmented", "ATT", or "Survey".}
}
\value{
\code{data.frame} containing estimated parameter values for a confounder with specified relative confounder strength to an observed covariate (or set of covariates), as well as the estimated bias from such an omitted confounder.
}
\description{
Returns parameter estimates for an omitted variable with specified relative confounding strength to an observed covariate (or set of covariates)
}
