% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_extreme_scenario}
\alias{calculate_extreme_scenario}
\title{Calculate extreme scenario}
\usage{
calculate_extreme_scenario(
  rho_w,
  weights,
  sigma2,
  correlations = c(0.25, 0.5, 0.9, 1)
)
}
\arguments{
\item{rho_w}{Estimated correlation between the estimated weights and the outcomes+}

\item{weights}{Vector of estimated weights}

\item{sigma2}{Variance of the outcomes}

\item{correlations}{A vector containing possible correlation values between the true weights and the outcomes}
}
\value{
A data frame that can be used to generate the extreme scenario plots
}
\description{
Helper function for running an extreme scenario analysis
}
