% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ces}
\alias{ces}
\title{Cooperative Election Survey}
\usage{
data('ces')
}
\description{
\itemize{
\item \code{pid}: party indicator, one of \code{c('Democrat', 'Republican', 'Independent', 'Other')}
\item \code{educ}: education level, one of \code{c('High School or Less', 'Some college', 'College', 'Post-grad')}
\item \code{age}: age of the individual
\item \code{age_buckets}: grouped age buckets, one of \code{c('18to35', '36to50', '51to64', 'Over65')}
\item \code{bornagain}: indicator for whether the individual identifies as born-again Christian, one of \code{c('Yes', 'No')}
\item \code{gender}: categorical gender, one of \code{c('Men', 'Women')}
\item \code{race}: categorical race, one of \code{c('White', 'Black', 'Hispanic', 'Asian', 'Other')}
\item \code{vvweight_post}: post-election verified voter weight for the individual
}
}
\details{
A dataset containing TODO (document the goal of the data)
}
\examples{
data('ces')
}
\references{
TODO (document where you got this from)
}
