% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{contour_plot}
\alias{contour_plot}
\title{Bias Contour Plots}
\usage{
contour_plot(
  varW,
  sigma2,
  killer_confounder,
  df_benchmark,
  benchmark = TRUE,
  shade = FALSE,
  shade_var = NULL,
  print = FALSE,
  binwidth = NULL
)
}
\arguments{
\item{varW}{Variance of the estimated weights}

\item{sigma2}{Estimated variance of the outcome (i.e., var(Y) for obervational setting; var(tau) for generalization setting)#'}

\item{killer_confounder}{Threshold for bias considered large enough to be a killer confounder. For example, if researchers are concerned about the bias large enough to reduce an estimated treatment effect to zero or change directional sign, set \code{killer_confounder} equal to the point estimate.}

\item{df_benchmark}{Data frame containing formal benchmarking results. The data.frame  must contain the columns \code{variable} (for the covariate name), \code{R2_benchmark}, and \code{rho_benchmark}.}

\item{benchmark}{Flag for whether or not to display benchmarking results (\code{benchmark = TRUE} if we want to add benchmarking results to plot, \code{benchamrk=FALSE} otherwise). If set to \code{TRUE}, \code{df_benchmark} must contain valid benchmarking results.}

\item{shade}{Flag for whether or not a specific benchmarking covariate (or set of benchmarked covariates) should be shaded a different color (\code{shade = TRUE} indicates that we want to highlight specific variables)}

\item{shade_var}{If \code{shade = TRUE}, this contains either a vector containing the variables we want to highlight}

\item{print}{If set to \code{TRUE}, the function will return a list with two elements: \code{plot} which contains the generated bias contour plot, and \code{data}, which provides the data.frame for generating the contour plot. If set to \code{FALSE}, the function will simply generate the bias contour plot. Default is set to \code{FALSE}.}
}
\value{
A ggplot2 object, containing the bias contour plot
}
\description{
Generates bias contour plots to aid with sensitivity analysis
}
