% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_utils.R
\name{contourPlotOptionsUI}
\alias{contourPlotOptionsUI}
\title{Custom plot options for contour plot}
\usage{
contourPlotOptionsUI(
  id,
  total_width = 12,
  height = 500,
  axis_text_size = 15,
  axis_title_size = 20,
  axis_line_width = 1,
  other_options = NULL
)
}
\description{
Custom plot options for contour plot
}
\keyword{internal}
