% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{extreme_scenario_plot}
\alias{extreme_scenario_plot}
\title{Extreme Scenario Plots}
\usage{
extreme_scenario_plot(
  rho_w,
  weights,
  sigma2,
  estimate,
  correlations = c(0.25, 0.5, 0.9, 1)
)
}
\arguments{
\item{rho_w}{Correlation between the estimated weights and the outcomes}

\item{weights}{Vector of estimated weights}

\item{sigma2}{Estimated variance of the outcome (i.e., stats::var(Y) for obervational setting; stats::var(tau) for generalization setting)#'}

\item{estimate}{Weighted estimate}

\item{correlations}{A vector containing possible correlation values between the true weights and the outcomes}
}
\value{
A ggplot2 object, genearting an extreme scenario analysis
}
\description{
Generates extreme scenario plots, with varying thresholds for the correlation between the true weights and the outcomes
}
