% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_rho}
\alias{generate_rho}
\title{Generate Correlation}
\usage{
generate_rho(rho_w, k, R2)
}
\arguments{
\item{rho_w}{Estimated correlation between the estimated weights and the outcomes}

\item{k}{Scaling factor, denoting how many times larger the correlation between the true weights and the outcomes are, relative to the correlation between the estimated weights and the outcomes}

\item{R2}{R^2 measure for how much variation in the true weights is explained by the error term, must be bound on the range \code{[0,1)}}
}
\value{
Correlation between the error and the weights, based on a user-input \code{R2} value and \code{stats::cor(w,Y)}
}
\description{
Returns the correlation between the error and the weights, using a user-inputted \code{R2} value, the correlation between the estimated weights and outcomes, and a scaling factor
}
