% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{psid_controls}
\alias{psid_controls}
\title{Lalonde Experimental Data (Dehejia and Wahba 1999 Subset)}
\usage{
data('psid_controls')
}
\value{
A data frame with 2490 observations and 10 variables
}
\description{
\itemize{
\item \code{treat}: Treatment assignment status (1 for treated, 0 for control)
\item \code{age}: Age of the individual
\item \code{education}: Years of education completed
\item \code{black}: Indicator for whether the individual is black
\item \code{hispanic}: Indicator for whether the individual is Hispanic
\item \code{married}: Indicator for whether the individual is married
\item \code{nodegree}: Indicator for whether the individual has no degree
\item \code{re74}: Earnings in 1974
\item \code{re75}: Earnings in 1975
\item \code{re78}: Earnings in 1978 (outcome variable)
}
}
\details{
Data from the National Supported Work Demonstration, a job training program
studied by Lalonde (1986). The dataset includes both treated and control
groups, with various demographic and earnings characteristics. This extract
was created Lalonde and distributed by Dehejia and Wahba (1999).
}
\examples{
data('psid_controls')
}
\references{
"Causal Effects in Non-Experimental Studies: Reevaluating the Evaluation of
Training Programs," Journal of the American Statistical Association, Vol. 94,
No. 448 (December 1999), pp. 1053-1062.

"Propensity Score Matching Methods for Non-Experimental Causal Studies,"
Review of Economics and Statistics, Vol. 84, (February 2002), pp. 151-161.
}
