% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run Shiny SenseWeight application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments to pass to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\value{
A shiny application
}
\description{
Run Shiny SenseWeight application
}
\examples{
if (interactive()) {
 run_app()
}
}
