% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senseweight-package.R
\docType{package}
\name{senseweight-package}
\alias{senseweight}
\alias{senseweight-package}
\title{senseweight: Sensitivity Analysis for Weighted Estimators}
\description{
This package provides functions and an interactive Shiny app to conduct rigorous sensitivity analyses for weighted estimators.
}
\author{
\strong{Maintainer}: Melody Huang \email{melody.huang@yale.edu}

Authors:
\itemize{
  \item Tiffany Tang \email{ttang4@nd.edu}
}

}
\keyword{internal}
