% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summarize_sensitivity}
\alias{summarize_sensitivity}
\title{Sensitivity Summary}
\usage{
summarize_sensitivity(
  weights,
  Y,
  Z,
  b_star = 0,
  estimate = NULL,
  SE = NULL,
  unweighted = NULL,
  sigma2 = NULL,
  estimand = "ATT",
  pretty = FALSE,
  dependent_var = NULL,
  model = NULL,
  outcome_function = NULL,
  svy_srs = NULL,
  svy_wt = NULL
)
}
\arguments{
\item{weights}{Vector of estimated weights}

\item{Y}{Outcome of interest}

\item{Z}{Treatment assignment}

\item{b_star}{TODO (document me!)}

\item{estimate}{(Optional) Weighted point estimate. If not specified, function will automatically generate the weighted estimator, given the inputs in \code{Y} and \code{weights}}

\item{SE}{(Optional) Standard error associated with the weighted point estimate}

\item{unweighted}{(Optional) Unweighted point estimate.}

\item{sigma2}{(Optional) Variance of outcomes or individual-level treatment effect in PATE case. In the case of a PATE estimator, if not specified, function will automatically estimate an upper bound for the variance of the individual-level treatment effect.}

\item{estimand}{Specifies estimand; possible parameters include "ATT", "PATE", or "Survey"}

\item{pretty}{If set to \code{TRUE}, will return a Kable table. If set to \code{FALSE}, will return a data.frame.}

\item{dependent_var}{TODO (document me!)}

\item{model}{TODO (document me!)}

\item{outcome_function}{TODO (document me!)}

\item{svy_srs}{TODO (document me!)}

\item{svy_wt}{TODO (document me!)}
}
\value{
Sensitivity summary
}
\description{
Returns a data.frame or Kable table with summary measures of sensitivity
}
\examples{
# TODO
}
