% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_utils.R
\name{updateWeightingOptions}
\alias{updateWeightingOptions}
\title{Update weighting options when data changes}
\usage{
updateWeightingOptions(id, data, is_default_data, Y, treatment, analysis_type)
}
\description{
Update weighting options when data changes
}
\keyword{internal}
