% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_utils.R
\name{weightingOptionsUI}
\alias{weightingOptionsUI}
\title{UI for weighting options}
\usage{
weightingOptionsUI(
  id,
  weighting_methods,
  weighting_method_tooltip = NULL,
  weighting_vars_manual_tooltip = NULL,
  weighting_vars_tooltip = NULL,
  picker_width = "100\%"
)
}
\description{
UI for weighting options
}
\keyword{internal}
