% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark an instance}
\usage{
benchmark(
  omit,
  weights,
  data,
  sigma2,
  weighting_method = "ebal",
  weight_max = Inf,
  estimand = "ATT"
)
}
\arguments{
\item{omit}{Variable to omit}

\item{weights}{Vector of estimated weights}

\item{data}{data.frame containing outcomes and covariate infromation}

\item{weighting_method}{Weighting method. Supports weighting methods from the package \code{WeightIt}.}

\item{estimand}{Specifies estimand; possible parameters include "ATT", "PATE", or "Survey"}
}
\value{
Robustness value for a specified proportion change
}
\description{
Returns the benchmarking results for a single covariate (or a single group of covariates)
}
