% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_functions.R
\name{benchmark_survey}
\alias{benchmark_survey}
\title{Benchmark (for survey weights)}
\usage{
benchmark_survey(
  drop_variable,
  formula,
  y,
  design,
  target,
  weights_orig,
  calfun_choice = "raking"
)
}
\arguments{
\item{drop_variable}{Variable to benchmark}

\item{formula}{Raking formula}

\item{y}{Response variable}

\item{design}{Survey object, containing the survey sample being re-weighted}

\item{target}{Survey object, containing the population the survey sample is being re-weighted to}

\item{weights_orig}{Weights used for analysis}

\item{calfun_choice}{Weighting method (default to raking)}
}
\value{
Benchmarking results for a variable (or subset of variables)
}
\description{
Returns benchmarking results for survey weighting
}
