% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_functions.R
\name{benchmark_survey}
\alias{benchmark_survey}
\title{Benchmark (for survey weights)}
\usage{
benchmark_survey(
  omit,
  formula,
  weights,
  pop_svy,
  sample_svy,
  Y,
  weighting_method
)
}
\arguments{
\item{omit}{Variable to benchmark}

\item{formula}{Raking formula}

\item{weights}{A vector, containing the estimated survey weights}

\item{pop_svy}{Survey object, containing the population the survey sample is being re-weighted to}

\item{sample_svy}{Survey object, containing the survey sample being re-weighted}

\item{Y}{outcome of interest}

\item{weighting_method}{Weighting method (default to raking)}
}
\value{
Benchmarking results for a variable (or subset of variables)
}
\description{
Returns benchmarking results for survey weighting
}
\examples{
# TODO
}
