% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_targets}
\alias{create_targets}
\title{Function for creating targets from auxiliary information and formula}
\usage{
create_targets(target_design, target_formula)
}
\description{
Function for creating targets from auxiliary information and formula
}
\keyword{internal}
