% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{robustness_value}
\alias{robustness_value}
\title{Robustness Value}
\usage{
robustness_value(q = 1, estimate, sigma2, weights)
}
\arguments{
\item{q}{Proportion of the point estimate that the bias is equal to (i.e., at \code{q=1}, the bias will be as large as the point estimate, indiating the bias is large enough to reduce the estimate to zero)}

\item{estimate}{Weighted estimate}

\item{sigma2}{Estimated variance of the outcome (i.e., var(Y) for obervational setting; var(tau) for generalization setting)}

\item{weights}{Vector of estimated weights}
}
\value{
Robustness value for a specified proportion change
}
\description{
Returns the estimated robustness value, for a specified proportion change
}
