% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{run_benchmarking}
\alias{run_benchmarking}
\title{Run Formal Benchmarking}
\usage{
run_benchmarking(
  weighting_vars,
  benchmark_vars = "all",
  data,
  treatment,
  outcome,
  selection,
  weighting_method = "ebal",
  weight_max = Inf,
  estimate,
  RV,
  sigma2,
  estimand = "ATT"
)
}
\arguments{
\item{weighting_vars}{Vector of variables to use in the weights estimation}

\item{benchmark_vars}{Vector of variables to benchmark parameters for. If \code{benchmark_vars = 'all'}, benchmarking will be run across all variables included in the weights. If not set to \code{all}, benchmarking will be conducted across the covariates included in the vector.}

\item{data}{A data.frame containing the observed covariates included in the weights; must include variables specified in weighting_vars}

\item{treatment}{Denotes which variable is the treatment variable}

\item{outcome}{Denotes which variable is the outcome variable}

\item{selection}{Denotes which variable is the selection variable}

\item{weighting_method}{Weighting method. Supports weighting methods from the package \code{WeightIt}.}

\item{weight_max}{Maximum weight to trim at. Default set to \code{Inf}.}

\item{estimate}{Weighted estimate}

\item{RV}{Robustness Value}

\item{sigma2}{If \code{estimand = "PATE"}, \code{sigma2} must specify the bound on treatment effect heterogeneity. For the other two estimands, the function will automatically calculate the sample variance across the control units, or the survey sample.}

\item{estimand}{Specifies estimand; possible parameters include "ATT", "PATE", or "Survey"}
}
\value{
data.frame containing the benchmarked parameter values, the estimated bias, MRCS, and minimum \code{k_sigma} and \code{k_rho} values for a killer confounder for the set of pre-specified covariates.
}
\description{
Wrapper function to run formal benchmarking on a set of pre-specified covariates. Returns a data.frame containing the benchmarked parameter values, the estimated bias, MRCS, and minimum \code{k_sigma} and \code{k_rho} values for a killer confounder.
}
\examples{
# TODO
}
