% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senseweight-package.R
\docType{package}
\name{senseweight-package}
\alias{senseweight}
\alias{senseweight-package}
\title{senseweight: Sensitivity Analysis for Weighted Estimators}
\description{
Provides tools to conduct rigorous sensitivity analyses for weighted estimators, based on (list papers). TODO Add two more sentences that describes specifics of the package.
}
\author{
\strong{Maintainer}: Melody Huang \email{melody.huang@yale.edu}

Authors:
\itemize{
  \item Tiffany Tang \email{ttang4@nd.edu}
}

}
\keyword{internal}
