% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summarize_sensitivity}
\alias{summarize_sensitivity}
\title{Sensitivity Summary}
\usage{
summarize_sensitivity(
  weights,
  Y,
  Z,
  q = 1,
  estimate = NULL,
  unweighted = NULL,
  sigma2 = NULL,
  estimand = "ATT"
)
}
\arguments{
\item{weights}{Vector of estimated weights}

\item{Y}{Outcome of interest}

\item{Z}{Treatment assignment}

\item{q}{Proportion to evaluate roustness value at. Default at \code{q=1}}

\item{estimate}{(Optional) Weighted point estimate. If not specified, function will automatically generate the weighted estimator, given the inputs in \code{Y} and \code{weights}}

\item{unweighted}{(Optional) Unweighted point estimate.}

\item{sigma2}{(Optional) Variance of outcomes or individual-level treatment effect in PATE case. In the case of a PATE estimator, if not specified, function will automatically estimate an upper bound for the variance of the individual-level treatment effect.}

\item{estimand}{Specifies estimand; possible parameters include "ATT", "PATE", or "Survey"}
}
\value{
Sensitivity summary
}
\description{
Returns a Kable table with summary measures of sensitivity
}
