% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_functions.R
\name{summarize_sensitivity_survey}
\alias{summarize_sensitivity_survey}
\title{Sensitivity Summary (for survey weights)}
\usage{
summarize_sensitivity_survey(svy_srs, svy_wt, weights, varY, b_star = 0)
}
\arguments{
\item{svy_srs}{Survey object, containing the unweighted survey}

\item{svy_wt}{Survey object, containing the weighted survey}

\item{weights}{A vector, containing the estimated survey weights}

\item{varY}{variance of the outcome}

\item{b_star}{Killer confounder threshold, default set to be zero}
}
\description{
Returns a data.frame or Kable table with summary measures of sensitivity; helper function for main summary function, and allows users to directly input a survey object and a design object
}
