% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_functions.R
\name{summarize_sensitivity_survey}
\alias{summarize_sensitivity_survey}
\title{Sensitivity Summary (for survey weights)}
\usage{
summarize_sensitivity_survey(
  dependent_var,
  model,
  outcome_function,
  svy_srs = NULL,
  Y = NULL,
  b_star = 0
)
}
\arguments{
\item{dependent_var}{String that provides the name of the outcome variable of interest}

\item{model}{Weighting model (i.e., a calibrate object)}

\item{outcome_function}{}
}
\description{
Returns a data.frame or Kable table with summary measures of sensitivity; helper function for main summary function, and allows users to directly input a survey object and a design object
}
